Coverage increase by utilizing the CONNECT method

+++ test.sh
@@ -11,7 +11,22 @@ origin_port=8080
 injector_tcp_port=8005
 proxy_tcp_port=5678
 HTTP_OK=200
+HTTP_MOVED=301
+HTTP_FOUND=302
 HTTP_BAD_GATEWAY=502
+
+trap cleanup SIGINT SIGTERM EXIT
+
+all_jobs=()
+function cleanup {
+    kill -SIGINT ${all_jobs[*]} 2>/dev/null || true
+}
+
+function element_in {
+    local e
+    for e in "${@:2}"; do [[ "$e" == "$1" ]] && return 0; done
+    return 1
+}
 
 function seconds_since_epoch {
     date +'%s'
@@ -29,31 +44,25 @@ function prepend {
 }
 
 function do_curl {
-    port=$1
-    expect=$2
-    host="$origin_addr:$origin_port"
-
-    code=$(curl -x localhost:$port $host -o /dev/null -w "%{http_code}" --silent --show-error)
-    
-    if [ "$code" != "$expect" ]; then
-        echo "$(now) $1 HTTP error code: $code (expected $expect)"
-        return 1
-    fi
+    local proxy=localhost:$1
+    local host=$2
+    local code=$(curl -x $proxy $host -o /dev/null -w "%{http_code}" --silent --show-error)
+    element_in $code "${@:3}" && return 0 || return 1
 }
 
 function test_n {
-    n=$1
-    expect=$2
+    local n=$1
+    local host=$2
     echo "$(now) Testing with $n jobs"
 
-    pids=()
+    local pids=()
     
     for ((i=0;i<$n;i++)); do
-        do_curl $proxy_tcp_port $expect &
+        do_curl $proxy_tcp_port $host "${@:3}" &
         pids+=("$!")
     done
     
-    r=0
+    local r=0
     for p in "${pids[@]}"; do
         if [ "$r" == "0" ]; then
             wait $p || r=1
@@ -68,24 +77,25 @@ function test_n {
 if [ "$origin_addr" == "localhost" ]; then
     ./test_server & # listens on origin_port
     server_pid=$!
+    all_jobs+=("$server_pid")
 fi
 
 echo "$(now) Starting injector."
 $unbuf ./injector -p 7000 2> >(prepend "Ie") 1> >(prepend "Io") &
-i_pid=$!
+all_jobs+=("$!")
 
 # Make sure injector starts properly.
 sleep 2
 
 echo "$(now) Testing curl directly to the server."
-do_curl $origin_port $HTTP_OK
+do_curl $origin_port $origin_addr:$origin_port $HTTP_OK
 
 echo "$(now) Testing curl to injector."
-do_curl $injector_tcp_port $HTTP_OK
+do_curl $injector_tcp_port $origin_addr:$origin_port $HTTP_OK
 
 echo "$(now) Starting injector helper."
 $unbuf ./injector_helper -i 127.0.0.1:7000 2> >(prepend "He") > >(prepend "Ho") &
-h_pid=$!
+all_jobs+=("$!")
 
 # Wait for the injector helper to perform a test on the injector nedpoint.
 sleep 2
@@ -93,26 +103,32 @@ sleep 2
 echo "$(now) Testing curl to injector_helper."
 r=0
 for i in 1 8 16; do
-    if ! test_n $i $HTTP_OK; then
+    if ! test_n $i $origin_addr:$origin_port $HTTP_OK; then
         r=1; break;
     fi
 done
 
 if [ "$r" == "0" ]; then
+    echo "$(now) Testing HTTPS forwarding."
+    if ! do_curl $proxy_tcp_port https://google.com $HTTP_OK $HTTP_FOUND $HTTP_MOVED; then
+        r=2
+    fi
+fi
+
+if [ "$r" == "0" -a -n "$server_pid" ]; then
+    echo "$(now) Testing fast failure response."
     # Kill the origin and try to connect to it. It shouldn't take long for the
     # injector and proxy to find out the origin is unreachable.
     kill -SIGINT $server_pid
 
     start_time=$(seconds_since_epoch)
-    test_n 2 $HTTP_BAD_GATEWAY
+    test_n 2 $origin_addr:$origin_port $HTTP_BAD_GATEWAY
     end_time=$(seconds_since_epoch)
 
     if [ $(($end_time - $start_time)) -ge 5 ]; then
         r=2
     fi
 fi
-
-kill -SIGINT $i_pid $h_pid 2>/dev/null || true
 
 echo "$(now) DONE"
